Prepared for no slicing

+++ testFold.py
@@ -15,20 +15,20 @@ from ModeTonicEstimation import Evaluator as ev
 # I/O
 #input_num = int(sys.argv[1])-1
 
-base_dir = '../../experiments/raga/'
+base_dir = '../../experiments/raag-recognition/'
 data_dir = os.path.join(base_dir,'data')
 experiments_dir = os.path.join(base_dir, 'experiments')
 modes = fo.getModeNames(data_dir)
 
 evaluator = ev.Evaluator()
-che = Chordia.Chordia(step_size=10, smooth_factor=15, chunk_size=120, 
+che = Chordia.Chordia(step_size=10, smooth_factor=15, chunk_size=0, 
                       threshold=0.5, overlap=0, frame_rate=128.0/44100)
 
 
 # indexing
 n_exp = 20
 n_folds = 14
-for input_num in range(0, 240):
+for input_num in range(0, n_exp*n_folds):
   mode_idx = np.unravel_index(input_num, [n_exp, n_folds])
 
   trainFolder = os.path.join(experiments_dir, 'exp' + str(mode_idx[0]), 
@@ -36,32 +36,34 @@ for input_num in range(0, 240):
   resultFile = os.path.join(experiments_dir, 'exp' + str(mode_idx[0]), 
                             'fold' + str(mode_idx[1]), 'resultsNoSample.json')
 
-
-  # create load the raag from the experiment & fold corresponding to the input index
-  foldFile = os.path.join(experiments_dir, 'exp' + str(mode_idx[0]), 'folds.json')
-
-
-  with open(foldFile) as f:
-      folds = json.load(f)
-          
-  makamTestFiles = folds['fold'+str(mode_idx[1])]['test']
+  if os.path.isfile(resultFile):
+    print resultFile + ' exists!'
+  else:
+    # create load the raag from the experiment & fold corresponding to the input index
+    foldFile = os.path.join(experiments_dir, 'exp' + str(mode_idx[0]), 'folds.json')
 
 
-  # Raag Recognition
-  results = []
-  for rec in makamTestFiles:
-      rec['pitch'] = np.loadtxt(rec['file'])[:,1]
-      res = che.estimate(rec['pitch'], mode_names=modes, mode_dir=trainFolder, 
-                         est_mode=True, distance_method='bhat',
-                         metric='pcd', tonic_freq=rec['tonic'], 
-                         equalSamplePerMode=False)[0]
+    with open(foldFile) as f:
+        folds = json.load(f)
+            
+    makamTestFiles = folds['fold'+str(mode_idx[1])]['test']
 
-      # evaluate
-      results.append(evaluator.mode_evaluate(rec['file'], res, rec['mode']))
 
-  print('exp' + str(mode_idx[0]) + '_fold' + str(mode_idx[1]) + ":" 
-        " " + str(100*np.mean([r['mode_eval'] for r in results])) + '%')
+    # Raag Recognition
+    results = []
+    for rec in makamTestFiles:
+        rec['pitch'] = np.loadtxt(rec['file'])[:,1]
+        res = che.estimate(rec['pitch'], mode_names=modes, mode_dir=trainFolder, 
+                           est_mode=True, distance_method='bhat',
+                           metric='pcd', tonic_freq=rec['tonic'], 
+                           equalSamplePerMode=False)[0]
 
-  with open(resultFile, 'w') as r:
-      json.dump(results, r, indent=4)
+        # evaluate
+        results.append(evaluator.mode_evaluate(rec['file'], res, rec['mode']))
+
+    print('exp' + str(mode_idx[0]) + '_fold' + str(mode_idx[1]) + ":" 
+          " " + str(100*np.mean([r['mode_eval'] for r in results])) + '%')
+
+    with open(resultFile, 'w') as r:
+        json.dump(results, r, indent=4)
